+++ Assets/MyGameAssets/Scripts/System/ScoreSetter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0eb7ca49766cd46738a6c93fb624bac3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
