+++ Assets/MyGameAssets/Scripts/Title.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9ee46c6ad07de4917b35ff6cfe52d7a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
