+++ Assets/MyGameAssets/Scripts/Title/ConfigButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e75220dd2251a4d5d8b9e31312abf338
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
