+++ Assets/MyGameAssets/Scripts/Title/LanguageSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93c60a7eb96b448e9908f757ff345fc8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
